added first invoice parsing RegEx

+++ process_upwork_statements.py
@@ -1,6 +1,7 @@
 import os
 import zipfile
 import shutil
+import requests
 
 
 import io
@@ -10,6 +11,25 @@ from pdfminer.converter import TextConverter
 from pdfminer.layout import LAParams
 from pdfminer.pdfpage import PDFPage
 from pdfminer.high_level import extract_text
+
+
+def process_first_invoice(invoice_text):
+    '''
+    returns (invoic_id, date string, amount) on sucess,
+    None otherwise 
+    '''
+    import re
+
+    regex = r"INVOICE #[\s]+DATE[\s]+DUE DATE[\s]+TOTAL AMOUNT[\s]+TOTAL DUE[\s]+(\b[A-Z][\d]+)[\s]+\w{3}\s\d{2},\s\d{4}[\s]+(\w{3}\s\d{2},\s\d{4})[\s]+\$\d+\.\d{2}[\s]+\$(\d+\.\d{2})"
+    
+    matches = re.finditer(regex, invoice_text, re.MULTILINE)
+    if matches == None:
+        return None
+    result = []
+    for match in matches:
+        result.append(match.group(1,2,3))
+    return result[0]
+    
 
 
 def convert_pdf_to_txt(path):
@@ -38,12 +58,21 @@ def convert_pdf_to_txt(path):
     return text
 
 
+def get_usd_rate_on_data(date_string):
+    r = requests.get("https://www.nbrb.by/api/exrates/rates/usd?parammode=2&{}".format(date_string))
+    if r.status_code == requests.codes.ok:
+        return r.json()
+    else: 
+        return None
+
+
 source_dir = os.path.dirname(os.path.realpath(__file__))
 dir_name = os.path.join(source_dir, 'unzipped')
 print('unzippiing into {}'.format(dir_name))
 extension = ".zip"
 
-shutil.rmtree(dir_name)
+if os.path.exists(dir_name):
+    shutil.rmtree(dir_name)
 if not os.path.exists(dir_name):
     os.mkdir(dir_name)
 # os.chdir(dir_name) # change directory from working dir to dir with files
@@ -56,7 +85,11 @@ for item in os.listdir(source_dir):  # loop through items in dir
         zip_ref.extractall(dir_name)  # extract file to dir
         zip_ref.close()  # close file
 
-for pdfFile in os.listdir(dir_name):
+for pdfFile in os.listdir(dir_name)[0:1]:
     print('-------')
-    print(extract_text(os.path.join(dir_name, pdfFile)))
+    invoice_text = extract_text(os.path.join(dir_name, pdfFile))
     print('-------')
+    print(process_first_invoice(invoice_text))
+
+
+print(get_usd_rate_on_data('2021-11-01'))
